Add credential file
The file provides a authentication token to help create the type of user.
This is not the best way to do as it simple tells anyone the password to create an admin account but for demonstration is enough.
This file could be shifted to .gitignore or we could treat credentials of users as env variables and store them in local .env file